% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treatment.R
\name{assign_treatment}
\alias{assign_treatment}
\title{Random treatment assignment with blocks}
\usage{
assign_treatment(blocking, treatments = c("Treated", "Control"))
}
\arguments{
\item{blocking}{\code{\link{qb_blocking}} or \code{\link[scclust]{scclust}} object with
the blocked units.}

\item{treatments}{vector with treatment conditions. If unnamed vector, elements in the
vector indicates treatment conditions that will be assigned in equal
proportions within each block. If named integer vector, the element
names indicate treatment conditions which will be assigned in the
proportions indicated by the values. See below for details.}
}
\value{
Returns a factor with the assigned treatments.
}
\description{
Provided a blocking and treatment conditions, \code{assign_treatment} randomly
assigns treatments to the units in the sample so to best maintain the desired
proportion of treatments within each block.
}
\details{
Units not assigned to blocks will not be assigned treatments.
}
\examples{
# Example blocking
my_blocking <- qb_blocking(c("A", "A", "B", "C", "B",
                             "C", "C", "A", "B", "B"))

# Two treatment conditions in equal propositions
assign_treatment(my_blocking)

# Two treatment conditions in with twice as many controls as treated
assign_treatment(my_blocking, c("Treated" = 1, "Control" = 2))

# Three treatment conditions in equal propositions
assign_treatment(my_blocking, c("T1", "T2", "C"))

# Four treatment conditions in equal propositions
# This gives a warning because some blocks have less than four units
\dontrun{assign_treatment(my_blocking, c("T1", "T2", "T3", "C"))}

}
