% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb_blocking.R
\name{is.qb_blocking}
\alias{is.qb_blocking}
\title{Check qb_blocking object}
\usage{
is.qb_blocking(x)
}
\arguments{
\item{x}{object to check.}
}
\value{
Returns \code{TRUE} if \code{x} is a valid \code{\link{qb_blocking}}
   object, otherwise \code{FALSE}.
}
\description{
\code{is.qb_blocking} checks whether the provided object is a valid instance
of the \code{\link{qb_blocking}} class.
}
\details{
\code{is.qb_blocking} does not check whether the blocking itself is sensible
or whether it satisfies some set of constraints. See
\code{\link[scclust]{check_clustering}} for that functionality.
}
