% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickblock-package.R
\docType{package}
\name{quickblock-package}
\alias{quickblock-package}
\title{quickblock: Quick Threshold Blocking}
\description{
Provides functions for constructing near-optimal threshold blockings. The
package is made with large data sets in mind and derives blockings more than
an order of magnitude quicker than other methods.
}
\details{
See the package's website for more information:
\url{https://github.com/fsavje/quickblock}.

Bug reports and suggestions are greatly appreciated. They are best reported
here: \url{https://github.com/fsavje/quickblock/issues}.
}
\references{
Higgins, Michael J., Fredrik Sävje and Jasjeet S. Sekhon (2016),
\sQuote{Improving massive experiments with threshold blocking},
\emph{Proceedings of the National Academy of Sciences}, \bold{113:27}, 7369--7376.
\url{http://www.pnas.org/lookup/doi/10.1073/pnas.1510504113}
}
