% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickblock.R
\name{quickblock}
\alias{quickblock}
\title{Construct threshold blockings}
\usage{
quickblock(distances, size_constraint = 2L, caliper = NULL, ...)
}
\arguments{
\item{distances}{\code{\link[distances]{distances}} object or a numeric vector, matrix
or data frame. The argument describes the similarity of the units to be
blocked. It can either be preprocessed distance information from a
\code{\link[distances]{distances}} object (recommended), or raw
covariate data. When called with covariate data, Euclidean distances are
calculated unless otherwise specified.}

\item{size_constraint}{integer with the required number of units in each block.}

\item{caliper}{restrict the maximum within-block distance.}

\item{...}{additional parameters to be sent either to the
\code{\link[distances]{distances}} function when \code{distances} contains
covariate data, or to the underlying \code{\link[scclust]{sc_clustering}}
function.}
}
\value{
Returns a \code{\link{qb_blocking}} object with the blocks.
}
\description{
...
}
\examples{
# Construct example data
my_data <- data.frame(x1 = runif(100),
                      x2 = runif(100))

# Make distances
my_distances <- distances(my_data, dist_variables = c("x1", "x2"))

# Make blocking with at least two units in each block
quickblock(my_distances)

# Require at least three units in each block
quickblock(my_distances, size_constraint = 3)

# Impose caliper
quickblock(my_distances, caliper = 0.2)

# Call `quickblock` directly with covariate data (ie., not pre-calculating distances)
quickblock(my_data[c("x1", "x2")])

# Call `quickblock` directly with covariate data using Mahalanobis distances
quickblock(my_data[c("x1", "x2")], normalize = "mahalanobize")

}
\references{
Higgins, Michael J., Fredrik Sävje and Jasjeet S. Sekhon (2016),
\sQuote{Improving massive experiments with threshold blocking},
\emph{Proceedings of the National Academy of Sciences}, \bold{113:27}, 7369--7376.
\url{http://www.pnas.org/lookup/doi/10.1073/pnas.1510504113}
}
\seealso{
See \code{\link[scclust]{sc_clustering}} for the underlying function used
  to construct the blocks.
}
