% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickblock-package.R
\docType{package}
\name{quickblock-package}
\alias{quickblock-package}
\title{quickblock: Quick Threshold Blocking}
\description{
Provides functions for assigning treatments in randomized experiments using
near-optimal threshold blocking. The package is made with large data sets in
mind and derives blocks more than an order of magnitude quicker than other
methods.
}
\details{
See \code{\link{quickblock}} for the main blocking function.

See the package's website for more information:
\url{https://github.com/fsavje/quickblock}.

Bug reports and suggestions are greatly appreciated. They are best reported
here: \url{https://github.com/fsavje/quickblock/issues}.
}
\references{
Higgins, Michael J., Fredrik Sävje and Jasjeet S. Sekhon (2016),
   \sQuote{Improving massive experiments with threshold blocking},
   \emph{Proceedings of the National Academy of Sciences}, \bold{113:27}, 7369--7376.
}
